Long

+++ src/Netology/part2/Git/Main.java
@@ -2,7 +2,7 @@ package Netology.part2.Git;
 
 public class Main {
     public static void main(String[] args) {
-        SalesManager amount = new SalesManager(new int[]{20, 35, 10, 15, 8});
+        SalesManager amount = new SalesManager(new long[]{20, 35, 10, 15, 8});
         System.out.println(amount.max());
     }
 }
